requirements.txt
python-dotenv==1.0.0
requests==2.28.2
selenium==4.7.2
